Introduce TransformOpInterface
This interface generalizes the behavior of ops from the LinalgTransform
dialect. These ops now implement the interface instead of having the
interpreter switch based on the op kind. The interpreter is
significantly reduced and becomes extensible, supporting transformations
defined by other dialects than LinalgTransform.

This requires quite a bit of code motion in the repository, and some
further cleanups will be performed by separate commits.
